+++ addOns/network/src/main/java/org/zaproxy/addon/network/internal/handlers/TlsConfig.java
@@ -41,7 +41,6 @@ public class TlsConfig {
     private List<String> tlsProtocols;
     private boolean alpnEnabled;
     private List<String> applicationProtocols;
-    private String fallbackApplicationProtocol;
 
     /**
      * Constructs a {@code TlsConfig} with all the SSL/TLS protocol versions supported, with ALPN
@@ -50,11 +49,7 @@ public class TlsConfig {
      * <p>If no protocol is negotiated it falls back to HTTP/1.1.
      */
     public TlsConfig() {
-        this(
-                DEFAULT_PROTOCOLS,
-                true,
-                DEFAULT_APPLICATION_PROTOCOLS,
-                TlsUtils.APPLICATION_PROTOCOL_HTTP_1_1);
+        this(DEFAULT_PROTOCOLS, true, DEFAULT_APPLICATION_PROTOCOLS);
     }
 
     /**
@@ -70,19 +65,10 @@ public class TlsConfig {
      */
     public TlsConfig(
             List<String> tlsProtocols, boolean alpnEnabled, List<String> applicationProtocols) {
-        this(tlsProtocols, alpnEnabled, applicationProtocols, null);
-    }
-
-    private TlsConfig(
-            List<String> tlsProtocols,
-            boolean alpnEnabled,
-            List<String> applicationProtocols,
-            String fallbackApplicationProtocol) {
         this.tlsProtocols = TlsUtils.filterUnsupportedTlsProtocols(tlsProtocols);
         this.alpnEnabled = alpnEnabled;
         this.applicationProtocols =
                 TlsUtils.filterUnsupportedApplicationProtocols(applicationProtocols);
-        this.fallbackApplicationProtocol = fallbackApplicationProtocol;
     }
 
     /**
@@ -110,15 +96,6 @@ public class TlsConfig {
      */
     public List<String> getApplicationProtocols() {
         return applicationProtocols;
-    }
-
-    /**
-     * Gets the protocol to use if no protocol was negotiated (ALPN).
-     *
-     * @return the fallback protocol, or {@code null} if none.
-     */
-    public String getFallbackApplicationProtocol() {
-        return fallbackApplicationProtocol;
     }
 
     @Override